+++ HaulageApplication/Haulage/CustomerPages/CustomerPage.xaml.cs
@@ -2,11 +2,15 @@ using Haulage.BillviewModel;
 using SQLite;
 using System.Runtime.CompilerServices;
 using Haulage.CustomerPages;
+using Haulage.BaseClasses.Accounting;
 namespace Haulage;
 
 public partial class CustomerPage : ContentPage
 {
-	public CustomerPage()
+
+    private Customer _customer;
+    private Account _account;
+    public CustomerPage()
 	{
 		InitializeComponent();
 
@@ -19,7 +23,36 @@ public partial class CustomerPage : ContentPage
 
 
     }
-    
-       
-    
+
+    private async void CustomerTrackItemNavigationButtonClicked(object sender, EventArgs e)
+    {
+        await Navigation.PushAsync(new CustomerPages.TrackItemPage());
+
+    }
+
+    public CustomerPage(Customer customer, Account account)
+    {
+        InitializeComponent();
+        _customer = customer ?? throw new ArgumentNullException(nameof(customer));
+        _account = account ?? throw new ArgumentNullException(nameof(account));
+
+        // Initialize the entries
+        BillingDetailsEntry.Text = account.BillingDetails;
+        ContactDetailsEntry.Text = account.ContactDetails;
+    }
+
+    public void SaveButton_Click(object sender, EventArgs e)
+    {
+        if (_customer != null && _account != null)
+        {
+            _customer.ManageAccount(_account, BillingDetailsEntry.Text, ContactDetailsEntry.Text);
+            DisplayAlert("Success", "Account details updated successfully.", "OK");
+        }
+        else
+        {
+            DisplayAlert("Error", "Failed to update account details.", "OK");
+        }
+
+    }
+
 }
